+++ __tests__/expectSaga/wrapped-sagas.test.js
@@ -20,19 +20,13 @@ function* regularBackgroundSaga() {
 const createRegularSaga = effectCreator =>
   function* saga() {
     const task = yield effectCreator(regularBackgroundSaga);
-    return task.name;
-  };
-
-const createParallelSaga = effectCreator =>
-  function* saga() {
-    const [task] = yield [effectCreator(regularBackgroundSaga)];
-    return task.name;
+    return task.meta.name;
   };
 
 const createAllSaga = effectCreator =>
   function* saga() {
     const [task] = yield all([effectCreator(regularBackgroundSaga)]);
-    return task.name;
+    return task.meta.name;
   };
 
 const createRaceSaga = effectCreator =>
@@ -40,7 +34,7 @@ const createRaceSaga = effectCreator =>
     const { saga: task } = yield race({
       saga: effectCreator(regularBackgroundSaga),
     });
-    return task.name;
+    return task.meta.name;
   };
 
 function* cancelledBackgroundSaga() {
@@ -58,13 +52,6 @@ function* cancelledBackgroundSaga() {
 const createCancelSaga = effectCreator =>
   function* saga() {
     const task = yield effectCreator(cancelledBackgroundSaga);
-    yield call(delay, 50);
-    yield cancel(task);
-  };
-
-const createCancelParallelSaga = effectCreator =>
-  function* saga() {
-    const [task] = yield [effectCreator(cancelledBackgroundSaga)];
     yield call(delay, 50);
     yield cancel(task);
   };
@@ -101,24 +88,8 @@ test('wrapped sagas return a task with a name referring to the spawned saga', ()
     .run();
 });
 
-test('wrapped sagas return a task with a name referring to the parallel forked saga', () => {
-  const saga = createParallelSaga(fork);
-
-  return expectSaga(saga)
-    .returns('regularBackgroundSaga')
-    .run();
-});
-
 test('wrapped sagas return a task with a name referring to the `all` forked saga', () => {
   const saga = createAllSaga(fork);
-
-  return expectSaga(saga)
-    .returns('regularBackgroundSaga')
-    .run();
-});
-
-test('wrapped sagas return a task with a name referring to the parallel spawned saga', () => {
-  const saga = createParallelSaga(spawn);
 
   return expectSaga(saga)
     .returns('regularBackgroundSaga')
@@ -171,7 +142,7 @@ test('providers receive task with name referring to forked saga', () => {
   return expectSaga(saga)
     .provide({
       cancel(task, next) {
-        expect(task.name).toBe('cancelledBackgroundSaga');
+        expect(task.meta.name).toBe('cancelledBackgroundSaga');
         return next();
       },
     })
@@ -185,7 +156,7 @@ test('providers receive task with name referring to spawned saga', () => {
   return expectSaga(saga)
     .provide({
       cancel(task, next) {
-        expect(task.name).toBe('cancelledBackgroundSaga');
+        expect(task.meta.name).toBe('cancelledBackgroundSaga');
         return next();
       },
     })
@@ -193,24 +164,8 @@ test('providers receive task with name referring to spawned saga', () => {
     .run();
 });
 
-test('wrapped parallel forked sagas can detect cancellation', () => {
-  const saga = createCancelParallelSaga(fork);
-
-  return expectSaga(saga)
-    .put({ type: 'DONE', payload: 'cancelled' })
-    .run();
-});
-
 test('wrapped `all` forked sagas can detect cancellation', () => {
   const saga = createCancelAllSaga(fork);
-
-  return expectSaga(saga)
-    .put({ type: 'DONE', payload: 'cancelled' })
-    .run();
-});
-
-test('wrapped parallel spawned sagas can detect cancellation', () => {
-  const saga = createCancelParallelSaga(spawn);
 
   return expectSaga(saga)
     .put({ type: 'DONE', payload: 'cancelled' })
@@ -225,41 +180,13 @@ test('wrapped `all` spawned sagas can detect cancellation', () => {
     .run();
 });
 
-test('providers receive task with name referring to parallel forked saga', () => {
-  const saga = createCancelParallelSaga(fork);
-
-  return expectSaga(saga)
-    .provide({
-      cancel(task, next) {
-        expect(task.name).toBe('cancelledBackgroundSaga');
-        return next();
-      },
-    })
-    .put({ type: 'DONE', payload: 'cancelled' })
-    .run();
-});
-
 test('providers receive task with name referring to `all` forked saga', () => {
   const saga = createCancelAllSaga(fork);
 
   return expectSaga(saga)
     .provide({
       cancel(task, next) {
-        expect(task.name).toBe('cancelledBackgroundSaga');
-        return next();
-      },
-    })
-    .put({ type: 'DONE', payload: 'cancelled' })
-    .run();
-});
-
-test('providers receive task with name referring to parallel spawned saga', () => {
-  const saga = createCancelParallelSaga(spawn);
-
-  return expectSaga(saga)
-    .provide({
-      cancel(task, next) {
-        expect(task.name).toBe('cancelledBackgroundSaga');
+        expect(task.meta.name).toBe('cancelledBackgroundSaga');
         return next();
       },
     })
@@ -273,7 +200,7 @@ test('providers receive task with name referring to `all` spawned saga', () => {
   return expectSaga(saga)
     .provide({
       cancel(task, next) {
-        expect(task.name).toBe('cancelledBackgroundSaga');
+        expect(task.meta.name).toBe('cancelledBackgroundSaga');
         return next();
       },
     })
@@ -303,7 +230,7 @@ test('providers receive task with name referring to race forked saga', () => {
   return expectSaga(saga)
     .provide({
       cancel(task, next) {
-        expect(task.name).toBe('cancelledBackgroundSaga');
+        expect(task.meta.name).toBe('cancelledBackgroundSaga');
         return next();
       },
     })
@@ -317,7 +244,7 @@ test('providers receive task with name referring to race spawned saga', () => {
   return expectSaga(saga)
     .provide({
       cancel(task, next) {
-        expect(task.name).toBe('cancelledBackgroundSaga');
+        expect(task.meta.name).toBe('cancelledBackgroundSaga');
         return next();
       },
     })