+++ __tests__/testSaga/effects/delay.test.js
@@ -1,0 +1,13 @@
+// @flow
+import { delay } from 'redux-saga/effects';
+import testSaga from 'testSaga';
+
+function* mainSaga() {
+  yield delay(500);
+}
+
+test('handles delay effect', () => {
+  testSaga(mainSaga)
+    .next()
+    .delay(500);
+});