+++ src/testSaga/extractSagaHelperEffectName.js
@@ -1,4 +1,0 @@
-// @flow
-export default function extractSagaHelperEffectName(name: string): string {
-  return name.replace(/^(\w+)\(.*$/, '$1');
-}